+++ CHANGES.rst
@@ -10,3 +10,6 @@ Changelog
 - Add configurable option to set values on all content that would match the 
   executing rule.
   [instification]
+
+- Add configurable option to preserve modification date of updated objects.
+  [instification]